refactor(container): extract authentication guard providers
